Major changes to businesscardviewmodel
Still does not work as expected, but better I guess.

Still have no idea why my mock data is not being persisted in the businesscard list or why the businesscard list is not initalized, but is null.

Half tempted to hard code the mock data into the list explicitly so it at least shows up and delete it later.